Fix issue with GitHub links containing @
It happens that if a `<a href…>…</a>` link contains `@` in the visible text, GitHub interprets it as an email, changes its link to `mailto`, and changes the `rel` to `nofollow` (had to show source code of rendered page to figure that one out!).

So to fix this, I have to either not using `@aligator@ohai.social` as the link text (but instead something like just "Mastodon" as the link text)… or I have to trick GitHub by using `＠` aka `U+FF20` aka `FULLWIDTH COMMERCIAL AT` instead of the traditional `@` aka `U+0040`. I chose the latter.